Expose the “up to date messages” to build scans.

+++ subprojects/core/src/main/java/org/gradle/api/execution/internal/ExecuteTaskBuildOperationDetails.java
@@ -47,7 +47,7 @@ public class ExecuteTaskBuildOperationDetails implements ExecuteTaskBuildOperati
 
     @Override
     public long getTaskId() {
-        return System.identityHashCode(task);
+        return (long) System.identityHashCode(task);
     }
 
     @Override
